Created Readme.md (See Ext. Desc)
In an effort to honor the development cycle 
(and use an agile SCRUM-like approach), 
I will be documenting specifications so intended users/devs may 
contribute, validate, more accurately test, and reuse this program.

That being said, the specifications are currently burnt into
my head that I've worked on so far. This is insufficient to future
keepers of this project. I will be uploading:

- Product backlog
- Release backlogs (Incremental, Level by level)
- Sprint backlogs (Slowly completing each release backlog)

For the whole project. I'm not too concerned about time constraints
or burnout times (So why am I using SCRUM...). Oh well. 
Anyways, until I can document these, I'll stop coding for the sake of 
good practice and confusion.